+++ lib/new_relic/agent/agent_helpers/startup.rb
@@ -48,6 +48,7 @@ module NewRelic
         # Treatment of @started and env report is important to get right.
         def setup_and_start_agent(options = {})
           @started = true
+          @health_check.create_and_run_health_check_loop
           @harvester.mark_started
 
           unless in_resque_child_process?
@@ -129,6 +130,7 @@ module NewRelic
           if Agent.config[:monitor_mode]
             true
           else
+            NewRelic::Agent.agent.health_check.update_status(NewRelic::Agent::HealthCheck::AGENT_DISABLED)
             ::NewRelic::Agent.logger.warn('Agent configured not to send data in this environment.')
             false
           end
@@ -140,6 +142,7 @@ module NewRelic
           if Agent.config[:license_key] && Agent.config[:license_key].length > 0
             true
           else
+            NewRelic::Agent.agent.health_check.update_status(NewRelic::Agent::HealthCheck::MISSING_LICENSE_KEY)
             ::NewRelic::Agent.logger.warn('No license key found. ' +
               'This often means your newrelic.yml file was not found, or it lacks a section for the running ' \
               "environment, '#{NewRelic::Control.instance.env}'. You may also want to try linting your newrelic.yml " \
@@ -160,6 +163,7 @@ module NewRelic
           if key.length == 40
             true
           else
+            NewRelic::Agent.agent.health_check.update_status(NewRelic::Agent::HealthCheck::INVALID_LICENSE_KEY)
             ::NewRelic::Agent.logger.error("Invalid license key: #{key}")
             false
           end
@@ -180,6 +184,7 @@ module NewRelic
           end
 
           unless app_name_configured?
+            NewRelic::Agent.agent.health_check.update_status(NewRelic::Agent::HealthCheck::MISSING_APP_NAME)
             NewRelic::Agent.logger.error('No application name configured.',
               'The agent cannot start without at least one. Please check your ',
               'newrelic.yml and ensure that it is valid and has at least one ',